Move URL utilities to ./clientUtils/urls/
