+++ explorer/tsconfig.json
@@ -8,7 +8,6 @@
         { "path": "../clientUtils" },
         { "path": "../grapher" },
         { "path": "../gridLang" },
-        { "path": "../coreTable" },
-        { "path": "../urls" }
+        { "path": "../coreTable" }
     ]
 }
